requirements.txt
flask
flask-socketio
mysql-connector-python
opencv-python
ultralytics
dnspython
secure-smtplib
email-validator
itsdangerous
